Add files via upload
A python program written to create and test a neural network for solving a regression problem, with a focus on hyperparameter tuning, to obtain the most appropriate network. The database contains data related to different houses for sale in King County, USA, the aim of the project is to predict the house price based on these features.